# expression: implement vectorized evaluation for builtinAddDateStringRealSig