@wshwsh12 2019-11-17
It seems that the warnings number of vec function and non-vec function is different and the test fails. Please fix this CI. 
```
[2019-11-17T06:00:51.161Z] FAIL: builtin_time_vec_test.go:241: testEvaluatorSuite.TestVectorizedBuiltinTimeFunc
[2019-11-17T06:00:51.161Z] 
[2019-11-17T06:00:51.161Z] builtin_time_vec_test.go:242:
[2019-11-17T06:00:51.161Z]     testVectorizedBuiltinFunc(c, vecBuiltinTimeCases)
[2019-11-17T06:00:51.161Z] bench_test.go:1143:
[2019-11-17T06:00:51.161Z]     c.Assert(2*vecWarnCnt, Equals, totalWarns)
[2019-11-17T06:00:51.161Z] ... obtained uint16 = 0x0
[2019-11-17T06:00:51.161Z] ... expected uint16 = 0x383
```